Add files via upload
This is a surveillance and analysis system for a township.
Starts with face recognition, smart car parking system and it check the emission of the vehicle (by detecting the number plate of the car) by comparing  it with its registry date  - if its older  than 15 years then it is not allowed to enter. Also a vehicle will be allowed according to even-odd rule(odd vehicles will be allowed only on odd days) to keep an eye on pollution level.